Title: 특정 문자로 시작하는 문자열, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241008/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.c
@@ -31,7 +31,7 @@ int main() {
         }
     }
 
-    printf("%d %d",cnt ,strlen(a)/cnt);
+    printf("%d %.2f",cnt ,strlen(a)/(double)cnt);
